Refactorted to be a bit more OO and removed regex
The code is slightly longer now, but I think that it's more
maintainable. Each private function is a bit smaller and easy to
understand. Because of this refactoring I was able to split apart the
matching logic and assign the appropriate matcher when a NewGrepper is
called.

The second largest update is getting rid of the regexp. Although very
flexible. It took about half a second to MakeCompile, which accounted
for about a third of the total processing time of Search. Using the
strings package plus preselecting how to match resulted in an almost 50%
decrease in processing time.

Before:
BenchmarkSearch-8   	    1189	    859111 ns/op	  332009 B/op	    1879 allocs/op

After:
BenchmarkSearch-8   	    2467	    496737 ns/op	  228050 B/op	     805 allocs/op